Utils.isTimely should passthrough for Rules-Engine v2 (#14)
In current versions of the webapp, partner code is required to only emit tasks which are ready to be displayed to the user. Utils.isTimely is the mechanism used for this. With the rules-engine improvements in webapp 3.8, this responsibility shifts. Partner code should now emit all tasks and the webapp's rules-engine decides what to display. To this end - Utils.isTimely becomes a pass-through in nootils@4.x.