Remove test-wrapper and test-utils (#13)
The test-wrapper classes in medic-nootils make this package a weird combination of production functionality together with a test library. Since this test library is blocking the upgrade of the production code and since the test library has been deprecated in favour of medic-conf-test-harness, it is necessary to remove the test-wrapper and test-utils libraries from medic-nootils.